Owat.CallDisplay (PVE v0.0.1)

This module handles the call display registration.

Functions

approve!(id, user_id)

@spec approve!(String.t(), String.t()) :: :ok

Approves a call display.

session is the session of the user that approves the call display.

approved?(id)

@spec approved?(Owat.Model.CallDisplay.t() | String.t()) :: boolean()

Checks if a given call display is approved.

Will return false if it does not exist.

broadcast(display)

Callback implementation for Owat.CrudUtils.broadcast/1.

broadcast!(data)

@spec broadcast!([Owat.CallDisplay.Content.t()]) :: :ok

Broadcasts call display content to all call displays.

delete(id)

@spec delete(String.t()) :: {:ok, Owat.Model.CallDisplay.t()} | {:error, term()}

Deletes a call display.

get(id)

@spec get(String.t()) :: Owat.Model.CallDisplay.t() | nil

list()

@spec list() :: [Owat.Model.CallDisplay.t()]

Returns all call displays.

maybe_broadcast(arg)

maybe_delete(struct_or_changeset)

maybe_insert(changeset, preload_opts \\ [])

maybe_transact(transaction)

maybe_transact(transaction, change_to_broadcast)

maybe_update(changeset)

maybe_update(changeset, preload_opts)

maybe_update(state, changes, preload_opts)

register()

@spec register() :: Owat.Model.CallDisplay.t()

Registers a new Call Display resource.

This resource will be created unapproved without a name and with a random, probably unique 8 character code. The uniqueness of the code is not guaranteed.

transact(changesets)

update(id, name, group)

@spec update(String.t(), String.t(), String.t() | nil) ::
  {:ok, Owat.Model.CallDisplay.t()} | {:error, term()}

Updates a call display's name.