Owat.Channel (PVE v0.0.1)

Provides functionality for OWAT resource-, presence- and rpc-channels.

Usage

Generally only the Owat.Channel module has to be used. It imports the event/3 macro and subsequently uses Phoenix.Channel.

defmodule Owat.SomeChannel do
  use Owat.Channel

  event "event_name" do
    %{parameter: binding :: :string \ "default_val"} ->
      # Socket is available in the body
      reply({:ok, %{}})
      # but you can use the `ok/0` and `reply/1` macros
      ok() # same as reply({:ok, %{}})
  end
end

Wildcard topics

When defining a multi-topic channel events can be limited to a channel.

# in the socket module
channel "some:*", Owat.SomeChannel

# in the channel module
event "event_name" in "some:foo" do
  %{} -> reply({:ok, "only in some:foo"})
end

This also supports topic parts. The part variable is visible in the function body.

event "event_name" in "topic:" <> part do
  %{} -> reply({:ok, "you are in: " <> part})
end

Permission Checks

When not providing a do-block, all channels are protected by the :general_access permission. A provided do-block is expected to check permissions.

Any event can also be individually secured by a permission by setting the @permission attribute with the desired atom.

@permission :technical
event "event_name" do
  ok("You have permission")
end

The macro permission_gate/2 can be used for custom business-logic dependant permission checks.

Parameters

Events may or may not take parameters. When the event's do-block consists of -> expressions the event takes parameters. Currently only maps are supported as parameters. If the event does not take any parameters, any parameter given is ignored.

The same event may take multiple shapes of parameters. Parameter keys can be optional, but all required parameter keys must be present and no unspecified key may be given. When multiple paths overlap with optional keys the path with the least number of unset optional keys is taken.

Optional parameters

A parameter is optional when its binding is _ or prefixed with _.

event "event_name" do
  %{} -> ok()
  %{a: _a :: :string} -> ok()
  %{a: _a :: :string, b: _b :: :string} -> ok()
end

In this example when given no keys the first path is taken, because it has 0 optional keys not present. When b is given, the third path is taken and when only a is given the second path is given.

Default values

A parameter, both required and optional, can have a default value. This is done by using the \ operator.

event "event_name" do
  %{a: a :: :string \ "default_val"} -> ok()
end

Top bindings

The function body can access the entire event payload by binding it to a variable. This is done by using the := operator.

event "event_name" do
  %{
    a: a :: :string,
    b: _b :: :string
  } = data ->
    # if "b" is not passed the map will only have the key "a"
    do_something(data)
    reply({:ok, "You have permission"})
end

Type hints

The type of a parameter can be hinted using the :: operator. The type can be any atom supported by Owat.ZodTransformer, as well as {:enum, ["value"]}, Owat.Support.Color or a function that emits a datatype (like Owat.Model.Person.datamap()). See Owat.Model for details on how to define a datatype.

These type hints are used to generate Zod schemas for the events a channel offers and all channels referenced using the channel/2 macro will be exported to zod schema files.

Beware

Type hints do not provide any input validation

Known Limitations / To Do

It is not supported to have an event that has something other than a map as payload.

Type hints do not provide any input validation.

Functions

array(atom)
(macro)

data_for(model, opts \\ [], list)
(macro)

Defines the data producer for a resource channel.

This relies on the produce_data_map/1 functions of modules using Owat.Model.

The do block of the macro invocation shall produce the data to be broadcast as data event. When the produced data is a list each element will be mapped over the produce_data_map/1 function of the specified module. To activate the list mode, the module atom shall be placed in list brackets.

When the do block returns {:stop, response} the client will be evicted from the channel.

Examples

data_for Resource.Module do
  Resource.acquire_one_result()
end

data_for [Resource.Module] do
  Resource.acquire_many_results()
end

In case the do block needs to have access to the socket (for example to work discriminate against the topic) the :bind option can be used.

data_for Resource.Module, bind: socket do
  Resource.acquire_one_result(socket)
end

data_for_presence(model, opts \\ [], list)
(macro)

enum(values)
(macro)

event(name_topic, opts \\ [], list)
(macro)

This macro creates a handle_in/3 method for a Phoenix.Channel.

event "name" do
  %{} -> :response
end

# results in

def handle_in("name", params, %Phoenix.Socket{} = socket) do
  # Argument checks and optionally permission checks

  case params do
    # here the do-body
    # the params parameter is not visible
    %{} -> :response
  end
end

event name in "topic"

# corresponds to

def handle_in("name", params, %Phoenix.Socket{topic: "topic"} = socket)

find_best_path(arg1, arg2)

join_permission(atom)
(macro)

maybe_data_map(error)

ok(data \\ {:%{}, [], []})
(macro)

ok_or_error(data)
(macro)

parse_response_type(fields)

permission_gate(hint, list)
(macro)

Only execute a block when the given permission is held.

permission_gate :technical do
  ok("you have `technical` permission.")
end

# equivalent to

case Owat.Permissions.authorize(socket, :technical) do
  :ok ->
    ok("you have `technical` permission.")

  {:error, err} ->
    {:reply, {:error, %{code: err}}, socket}
end

presence(topic, datatype_fn, opts \\ [])
(macro)

Defines a presence channel.

A presence channel sends presence_state and presence_diff events.

Resource channels must always define a corresponding data_for_presence/3 block. Sending the initial data is already handled in this macro.

reply(data \\ {:%{}, [], []})
(macro)

resource(topic, model, block \\ [])
(macro)

Defines a resource channel.

A resource channel sends data events. An optional do block can be supplied to perform join checks.

Resource channels must always define a corresponding data_for/2 block. Sending the initial data is already handled in this macro.

Examples

Normal channel definition with resource module. Allows all clients to join.

resource "channel_name", Resource.Module

Channel definition that executes code on join which can refuse the socket to join.

resource "channel_name", Resource.Module do
  # return :ok to allow join or a {:error, response} tuple to refuse join.
end

rpc(topic, opts \\ [])
(macro)

Defines an RPC channel.

An RPC channel does not send the data event. A channel defined with RPC does not need a corresponding call to data_for/3

Example

rpc "channel_name"