Owat.Elga.ZPI (PVE v0.0.1)

Client module for ELGA Z-PI

Types

demographic_query_opt()

@type demographic_query_opt() ::
  :phonetic
  | :additional_names
  | :only_patients_alive
  | :all_patients
  | {:given, String.t()}
  | {:family, String.t()}
  | {:date_of_birth,
     Owat.Support.DateTime.partial_date()
     | %{
         high: Owat.Support.DateTime.partial_date(),
         low: Owat.Support.DateTime.partial_date()
       }}
  | {:gender, :M | :F | :UN}
  | {:country, String.t()}
  | {:postal_code, String.t()}
  | {:city, String.t()}
  | {:street_name, String.t()}
  | {:number, String.t()}

Query options for a Z-PI demographics query.

Required parameters are:

family

or

given

date_of_birth

The parameter gender and address-related parameters are always optional.

The date_of_birth can be either a value, a partial value, a range of values or a range of partial values.

Wildcard Parameters

Wildcard search enables searching by prefix. The following parameters support wildcard search, where W<N> means that at least N characters have to be provided before the wildcard:

given W4

family W4

postal_code W2

city W4

street W4

Special Search Opts

There are the following switches to tune searching by name:

phonetic - turns on phonetic matching of the given and family name.

additional_names - causes the search by given and family to be matched against all known names of the patient. Default is to only search by the current name.

only_patients_alive - only patients not marked as deceased will be returned.

all_patients - all identities in a linked group are searched.

identifier_query()

@type identifier_query() :: {String.t(), String.t()}

A {system, value} tuple.

zpi_data()

@type zpi_data() :: %{
  id: String.t(),
  name: [
    %{
      use: String.t() | nil,
      qualifier: String.t() | nil,
      valid_until: Owat.Support.DateTime.partial_date() | nil,
      prefix: String.t() | nil,
      given: String.t() | nil,
      family: String.t() | nil,
      suffix: String.t() | nil
    }
  ],
  address: [
    %{
      foreign_id: String.t(),
      care_of: String.t() | nil,
      additional_locator: String.t() | nil,
      valid_until: Owat.Support.DateTime.partial_date() | nil,
      country: String.t(),
      postal_code: String.t(),
      city: String.t(),
      street_name: String.t(),
      number: String.t(),
      stairway: String.t() | nil,
      floor: String.t() | nil,
      apartment: String.t() | nil
    }
  ],
  deceased: boolean(),
  date_of_birth: Owat.Support.DateTime.partial_date() | nil,
  date_of_death: Owat.Support.DateTime.partial_date() | nil,
  citizenship: String.t() | nil,
  multiple_birth: boolean(),
  multiple_birth_ord: integer() | nil,
  ssn: String.t() | nil,
  gender: String.t(),
  bpk: String.t() | nil
}

Demographics data available from Z-PI

Functions

parse_pdq_response(xml)

patient_demographic_query(hcp_assertion, query)

@spec patient_demographic_query(
  String.t(),
  identifier_query() | [demographic_query_opt()]
) ::
  {:ok, [zpi_data()]}
  | {:error,
     HTTPoison.Response.t()
     | HTTPoison.Error.t()
     | Owat.Elga.ZPI.Error.t()
     | {:multiple, [Owat.Elga.ZPI.Error.t()]}
     | :no_organization_oid
     | {:invalid_opt, term()}}

Performs a patient demographic query against the Z-PI.

Either performs a query by LivingSubject.id when given an identifier query tuple or a query by demographic data. For demographic query options see demographic_query_opt/0.

sfa_info(map)