Owat.Gin.Consult (PVE v0.0.1)

Module for consult operations.

Generally speaking a consult is a type of voucher that is obtained by submitting a request to the insurance provider. This voucher has to be shown when billing services to the provider as a proof of insurance.

There are complex rules as to when a consult may be obtained depending on the field of practice (often passed via a parameter field_code), on the case of treatment (often passed via a parameter case_code), whether the patient already has pending vouchers with other healthcare providers and so on.

For the purposes of this module a consult may have a state that can take one of the following values:

PENDING indicates that the consult is requested but the request is yet in-flight.

ACTIVE indicates that the consult is a valid, registered voucher

CANCELLED indicates that the consult is a registered voucher that has since been cancelled.

FAULT indicates that the consult was requested, but the voucher could not be issued due to either business or technical reasons

CLAIMS, EXTERN and SURROGATE_VOUCHER are special cases of FAULT where the underlying business reason is clearly specified and resolvable

ERROR indicates that the consult was requested, but it failed to be submitted to the upstream services for technical reasons (like network outage)

Happy-Path Lifecycle

A normal consult in the best conditions takes the following life-cycle:

Submission either via automatic/2 or manual/5

Best case it goes from PENDING directly to ACTIVE

Second best case an easy to resolve CLAIMS state

In this case the CLAIMS issue is resolved via supply_claim/2

Done. A nice consult in ACTIVE state.

May have its case code changed via change_case/2

May be cancelled via cancel/1

May have supplemental information that must be answered via a call to send_supplemental_answer/3

Error resolving

The most common error state besides CLAIMS is FAULT. A fault occurs mostly for business rule reasons and is resolved by calling retry/2 and supplying information to conform to the business rules.

Submission either via automatic/2 or manual/5

The consult enters FAULT state

Retry by using retry/2 until the state changes to ACTIVE

Functions

answer_supplemental_consult(request)

automatic(encounter)

@spec automatic(Owat.Model.Encounter.t()) ::
  {:ok, Owat.Model.Consult.t()} | {:error, term()}

Registers a consult using the automatic approach.

by_id(id)

@spec by_id(String.t()) :: Owat.Model.Consult.t() | nil

Returns a consult by its ID.

cancel(consult)

@spec cancel(Owat.Model.Consult.t()) ::
  {:ok, Owat.Model.Consult.t()} | {:error, term()}

Cancels an active consult.

change_case(consult, new_case_code)

@spec change_case(Owat.Model.Consult.t(), String.t()) ::
  {:ok, Owat.Model.Consult.t()} | {:error, term()}

Changes the case code of an active consult

extern_reason(consult, case_code)

@spec extern_reason(Owat.Model.Consult.t(), String.t()) ::
  {:ok, Owat.Model.Consult.t()} | {:error, term()}

Supplies the adequate case code to convey a reason for the encounter with an extern patient.

This function only takes consults in EXTERN state without vouchers.

manual(encounter, force_admin_card, field_code, case_code, opts)

@spec manual(Owat.Model.Encounter.t(), boolean(), String.t(), String.t(), [
  Owat.Gin.SS12.Kse.consult_opt()
]) :: {:ok, Owat.Model.Consult.t()} | {:error, term()}

Registers a consult manually.

md_relation(consult, new_relation)

@spec md_relation(Owat.Model.Consult.t(), String.t()) ::
  {:ok, Owat.Model.Consult.t()} | {:error, term()}

This function sets the md_relation flag of the person associated with the consult and then updates the consult.

When the new relation is EXTERN the consult will simply be set to extern. When the new relation is either PRIMARY or PRIMARY_TO_BE then a new consult will be registered. Setting the relation to nil is not supported.

This function accepts only consults in LOCKED (this state is now denoted by state FAULT with an error code of ZS-00015) or EXTERN state without vouchers.

provide_ecard_signature(reader, person, token)

@spec provide_ecard_signature(
  Owat.Gin.ECardReader.t(),
  Owat.Model.Person.t(),
  String.t()
) ::
  Owat.Model.ECardSignature.t()

Registeres an ecard signature with the system.

retry(consult, changes)

@spec retry(Owat.Model.Consult.t(), map()) ::
  {:ok, Owat.Model.Consult.t()} | {:error, term()}

Retry a faulted consult with other data.

Merges properties of changes into the data from the previous request.

retry_action(request)

@spec retry_action(Owat.Model.ConsultRequest.t()) ::
  {:ok, Owat.Model.Consult.t()} | {:error, term()}

Retries a specific action idempotently with its transaction ID.

send_supplemental_answer(consult, user_id, answer_value)

@spec send_supplemental_answer(Owat.Model.Consult.t(), String.t(), String.t()) ::
  :ok | {:error, term()}

Sends an answer for a supplemental info request.

supply_claim(consult, list)

@spec supply_claim(Owat.Model.Consult.t(), Keyword.t()) ::
  {:ok, Owat.Model.Consult.t()} | {:error, term()}

Supply claim information for a consult in claims-fault state.