Owat.Gin.SS12.Kse (PVE v0.0.1)

Consult Service

Types

consult_opt()

@type consult_opt() ::
  {:ssn, String.t()}
  | {:provider, String.t()}
  | {:claim, String.t()}
  | {:afterwards, %{date: Date.t(), reason: String.t()}}
  | {:offline, Date.t()}
  | {:surrogate, String.t()}

Creating a consult may take the following options:

:ssn - specifies the SSN to perform the consult for

:provider - specifies a provider in case of a previous claims fault

:claim - specifies a claims ID in case of a previous claims fault

:afterwards - a map specifying the :date of consult and the :reason for retroactive registering (this would be a tuple, but then it would not be json serializable)

:offline - specifies that this consult was recorded offline and is now being retried. This takes the date of the consult.

:surrogate - specifies a surrogate voucher code

When specifying both :offline and :afterwards the dates must be the same. If they are not, the last specified date wins.

search_criterion()

@type search_criterion() ::
  {:field_code, String.t()}
  | {:since, Date.t()}
  | {:until, Date.t()}
  | {:case_code, String.t()}
  | {:type, String.t()}
  | {:practice_id, String.t()}
  | {:ssn, String.t()}
  | {:provider_code, String.t()}

Functions

cancel_consult(dialog, transaction_id, consult_id, consult_version)

change_case_code(dialog, transaction_id, consult_id, consult_version, new_case_code)

get_consult_data(dialog, transaction_id, foreign_ids)

@spec get_consult_data(Owat.Gin.SS12.Dialog.t(), String.t(), [String.t()]) ::
  {:ok, map()} | {:fault, Owat.Gin.SS12.fault_info()}

get_consult_data(dialog, transaction_id, search_criteria, type, sort)

@spec get_consult_data(
  Owat.Gin.SS12.Dialog.t(),
  String.t(),
  [search_criterion()],
  :all | :modifiable | :cancelled,
  :date | :case | :field | :provider | :patient
) ::
  {:ok, map()}
  | {:fault, Owat.Gin.SS12.fault_info()}
  | {:error, :invalid_serach_criterion | :invalid_type | :invalid_sort}

reactivate_consult(dialog, transaction_id, consult_id, consult_version)

request!(operation, arguments)

request!(operation, arguments, transaction_id)

send_supplemental_answer(dialog, transaction_id, supplemental_id, answer_value)

submit_consult(dialog, transaction_id, card_token, case_code, field_code, opts \\ [])