Owat.Gin.Signing (PVE v0.0.1)

Signing supervisor

This module handles GINO-signing related functionality. It is meant to be started in the supervision tree and supervises the Signing-Jobs.

Types

admin_card_signing_problem()

@type admin_card_signing_problem() :: :no_pin | :declined | :no_reader_responded

Functions

change_vpnr(vpnr)

@spec change_vpnr(String.t()) :: :ok

Callback function to notify the supervisor that the VPNR has changed.

This will trigger a restart of all signing jobs.

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

ready?()

@spec ready?() :: boolean()

Returns true if the system should generally be ready to handle signatures.

This tests that all these conditions are met:

There is a process registered to the module atom

The process is alive

There are readers available

At least one of these readers is marked as admin

There is at least one admin card present on an admin card reader

sign_admin_card()

@spec sign_admin_card() ::
  {:ok, map(), String.t()} | {:error, admin_card_signing_problem()}

Obtains a VP-type signature with the admin card.

sign_with_admin_card(ssn)

@spec sign_with_admin_card(String.t()) ::
  {:ok, map(), String.t()}
  | {:error, :no_vpnr | :declined | :no_reader_responded}

Obtains a SV-type signature of a SSN with the admin card.

start_link(_)

update_readers(readers)

@spec update_readers([map()]) :: :ok

Callback function for the Detector-Job to notify the supervisor of the currently available readers.