Owat.Legacy.Windoc.Importer (PVE v0.0.1)

Imports legacy data from the Windoc system.

Types

import_data()

@type import_data() :: record() | [import_data()] | nil

An iodata-like structure of possibly nested lists of records.

Any nil values shall be ignored upon import. The structure of this data must follow the constraint that a depth first search of the lists results in correct import. That is every dependant item comes after its dependency.

original()

@type original() :: Owat.Legacy.Windoc.legacy_patient()

record()

@type record() :: %{
  :__type__ => module(),
  optional(:id) => String.t(),
  optional(term()) => term()
}

A single record to be insert into the DB.

Contains the schema module as :__type__. May contain an :id. If an :id is set the record must be inserted with this ID. If no ID is set, one must be auto-generated prior to insert. A record never contains timestamps. There must always be generated prior to insert.

Functions

import_all(patients, progress_reporter)

produce_address(arg1, person_id)

@spec produce_address(map(), String.t()) :: [Owat.Model.Residency.t()]

Finds the given address through multiple steps of increasingly inclusive fuzzy search on the meilisearch address book index.

The Street field is split into its street name and numbers parts using combined_with_street_number and the numbers are parsed using queries_compound_street_number. In the first step, the address book is checked for an exact match. If none can be found, the following fuzzy search attempts will be made until a satisfying result is returned:

Fuzzy search the number and exact matches on everything else

Fuzzy search the street name and exact matches on everything else

Fuzzy search the street name and number, making sure the provided number is a substring of the result.

Fuzzy search the city and number, making sure the provided number is a substring of the result.

Fuzzy search the city and use the result to fuzzy search the street name and number with an exact match on the found city, making sure teh number is a substring of the result.

If none of these steps yield a result, an empty list is returned. If a matching address has been found, it is returned as a list containing a single residency struct referencing the either an existing address book entry from the database or a newly created one.

produce_records(patient)

@spec produce_records(original()) :: import_data()

split_duplicates(legacy_patients)

@spec split_duplicates([original()]) :: {[original()], [{String.t(), [original()]}]}