Owat.MasterPatientIndex (PVE v0.0.1)

The master patient index GenServer

Types

create_opt()

@type create_opt() ::
  {:date_of_birth, Date.t()}
  | {:address, residency_arg()}
  | {:administrative_gender, String.t()}

residency_arg()

@type residency_arg() :: %{
  :country => String.t(),
  :postal_code => String.t(),
  :city => String.t(),
  :street_name => String.t(),
  :number => String.t(),
  optional(:stairway) => String.t(),
  optional(:floor) => String.t(),
  optional(:apartment) => String.t()
}

Functions

add_contact(person, source, system, use, before, value)

@spec add_contact(
  Owat.Model.Person.t(),
  String.t(),
  String.t(),
  String.t(),
  String.t(),
  integer()
) ::
  {:ok, Owat.Model.Person.t()} | {:error, term()}

add_residency(person, source, use, type, before, addr)

@spec add_residency(
  Owat.Model.Person.t(),
  String.t(),
  String.t(),
  String.t(),
  residency_arg(),
  integer()
) :: {:ok, Owat.Model.Residency.t()} | {:error, term()}

broadcast(contact)

Callback implementation for Owat.CrudUtils.broadcast/1.

by_id(id)

@spec by_id(String.t()) :: Owat.Model.Person.t() | nil

Obtains a person by their ID.

create_person(flags \\ %{})

Creates an empty person without any data besides an unique ID.

create_person(given, family, opts \\ [])

@spec create_person(String.t(), String.t(), [create_opt()]) :: Owat.Model.Person.t()

Creates a new person and immediately attaches name and life info with the given parameters.

create_person(source, use, type, system, value)

@spec create_person(String.t(), String.t(), String.t(), String.t(), String.t()) ::
  Owat.Model.Person.t()

Creates a new person and immediately attaches an identifier with the given parameters.

delete(person)

@spec delete(Owat.Model.Person.t()) :: :ok | {:error, :retention_policy | term()}

delete_contact(contact_id)

@spec delete_contact(String.t()) :: {:ok, Owat.Model.Person.t()} | {:error, term()}

delete_residency(residency_id)

@spec delete_residency(String.t()) :: {:ok, Owat.Model.Person.t()} | {:error, term()}

exists?(id)

@spec exists?(String.t()) :: boolean()

Check if a person with the given ID exists.

find_by_contact(system, use, value)

@spec find_by_contact(String.t(), String.t(), String.t()) ::
  Owat.Model.Person.t() | nil

Find at most one person with the given contact details.

find_by_identifier(system, value)

@spec find_by_identifier(String.t(), String.t()) :: Owat.Model.Person.t() | nil

find_by_identifier_or_create(source, use, type, system, value, data)

@spec find_by_identifier_or_create(
  String.t(),
  String.t(),
  String.t(),
  String.t(),
  String.t(),
  map()
) :: {:ok, Owat.Model.Person.t()} | {:error, term()}

find_residencies(source, use, data)

@spec find_residencies(String.t(), String.t(), residency_arg()) :: [
  Owat.Model.Residency.t()
]

get_by_contact(system, use, value)

@spec get_by_contact(String.t(), String.t() | :any, String.t()) :: [
  Owat.Model.Person.t()
]

Fetch possibly many persons with the given contact details.

The parameter use can take the value :any to not restrict the search by use.

get_by_identifier(system, value)

@spec get_by_identifier(String.t(), String.t()) :: [Owat.Model.Person.t()]

Fetch possibly many persons with the given identifier.

has_ssn?(pid)

@spec has_ssn?(String.t()) :: boolean()

list_contacts(pid)

@spec list_contacts(Owat.Model.Person.t()) :: [Owat.Model.ContactPoint.t()]

list_residencies(pid)

@spec list_residencies(String.t()) :: [Owat.Model.Residency.t()]

maybe_broadcast(arg)

maybe_delete(struct_or_changeset)

maybe_insert(changeset, preload_opts \\ [])

maybe_transact(transaction)

maybe_transact(transaction, change_to_broadcast)

maybe_update(changeset)

maybe_update(changeset, preload_opts)

maybe_update(state, changes, preload_opts)

merge(keep_id, absorb_id)

set_ssn(person, source, ssn)

@spec set_ssn(Owat.Model.Person.t(), String.t(), String.t()) ::
  {:ok, Owat.Model.Person.t()}
  | {:error,
     :no_ssn
     | :malformed
     | :checksum_mismatch
     | :has_ssn
     | :ssn_already_known
     | Owat.CrudUtils.validation_error()}

transact(changesets)

update_contact(contact_id, data)

@spec update_contact(String.t(), String.t()) ::
  {:ok, Owat.Model.Person.t()} | {:error, term()}

update_default_claim(person, claim)

@spec update_default_claim(
  Owat.Model.Person.t(),
  Owat.Model.Person.DefaultClaim.t() | :keep
) ::
  {:ok, Owat.Model.Person.t()} | {:error, term()}

update_identifier(person, source, use, type, system, value)

@spec update_identifier(
  Owat.Model.Person.t(),
  String.t(),
  String.t(),
  String.t(),
  String.t(),
  String.t()
) :: {:ok, Owat.Model.Person.t()} | {:error, term()}

update_md_relation(person, status)

@spec update_md_relation(
  Owat.Model.Person.t(),
  :keep | :PRIMARY | :PRIMARY_TO_BE | :EXTERN
) ::
  {:ok, Owat.Model.Person.t()} | {:error, term()}

update_person(person, source, data)

@spec update_person(Owat.Model.Person.t() | String.t(), String.t(), map()) ::
  {:ok, Owat.Model.Person.t()} | {:error, term()}

Updates a person with the given data.

update_person_from_sas_result(person, data)

update_person_from_zpi_data(zpi_data)

update_residency(person_id, residency_id, source, data)

@spec update_residency(String.t(), String.t(), String.t(), %{
  :address => residency_arg(),
  optional(:before) => String.t() | nil,
  optional(:use) => String.t()
}) :: {:ok, Owat.Model.Person.t()} | {:error, term()}

upsert_residencies(person, source, data)

@spec upsert_residencies(Owat.Model.Person.t(), String.t(), [
  %{address: residency_arg(), use: String.t()}
]) :: {:ok, Owat.Model.Person.t()} | {:error, term()}