Owat.Medication (PVE v0.0.1)

Functions relating to prescriptions.

Workflow of prescribing medications

A prescription may or may not start with a proposal request. A proposal request can be accepted by a practitioner to become a draft order. The draft order exists because a prescription may prescribe multiple products at once. Issuing the prescription will set all draft orders to active and set the group identifier.

Types

medication_order_input()

@type medication_order_input() ::
  %{
    id: String.t(),
    quantity: integer(),
    dosage: any(),
    note_to_provider: String.t()
  }
  | %{
      name: String.t(),
      dose: String.t(),
      kind: String.t(),
      quantity: integer(),
      dosage: any(),
      note_to_provider: String.t()
    }

medication_request_input()

@type medication_request_input() ::
  %{id: String.t(), quantity: integer()}
  | %{name: String.t(), dose: String.t(), kind: String.t(), quantity: integer()}

order_opt()

@type order_opt() ::
  {:dosage, map()} | {:note_to_provider, String.t()} | {:quantity, integer()}

Functions

accept_request(id, encounter_id, user_id, opts \\ [])

@spec accept_request(String.t(), String.t() | nil, String.t(), [order_opt()]) ::
  {:ok, Owat.Model.MedicationRequest.t()}
  | {:error, :not_found | :invalid_status | Owat.CrudUtils.validation_error()}

Accepts a request for a medication.

Sets the medication request to order intent. This means a prescription is about to be issued but since it may be composed of multiple requests this function does not issue it yet.

One may optionally supply a dosage when accepting a request.

Accepting a request requires the request to be in status active with intent proposal, otherwise the error :invalid_status will be returned.

accept_request_substituting(id, encounter_id, user_id, medication_id, opts \\ [])

@spec accept_request_substituting(
  String.t(),
  String.t() | nil,
  String.t(),
  String.t(),
  [order_opt()]
) ::
  {:ok, Owat.Model.MedicationRequest.t()}
  | {:error, :not_found | :invalid_status | Owat.CrudUtils.validation_error()}

Accepts a request in principle but overrides the medication.

broadcast(request)

Callback implementation for Owat.CrudUtils.broadcast/1.

by_id(id)

@spec by_id(String.t()) :: Owat.Model.MedicationRequest.t() | nil

by_id(id, person_id)

@spec by_id(String.t(), String.t()) :: Owat.Model.MedicationRequest.t() | nil

for_person(person_id)

@spec for_person(String.t()) :: [Owat.Model.MedicationRequest.t()]

issue_prescription(user_id, request_ids)

(Not implemented yet) Issues a prescription from the given request IDs.

mark_issued(id, encounter_id, user_id, opts \\ [])

@spec mark_issued(String.t(), String.t(), String.t(), [order_opt()]) ::
  {:ok, Owat.Model.MedicationRequest.t()}
  | {:error, :not_found | Owat.CrudUtils.validation_error()}

Marks a request as been prescribed.

maybe_broadcast(arg)

maybe_delete(struct_or_changeset)

maybe_insert(changeset, preload_opts \\ [])

maybe_transact(transaction)

maybe_transact(transaction, change_to_broadcast)

maybe_update(changeset)

maybe_update(changeset, preload_opts)

maybe_update(state, changes, preload_opts)

order(person_id, encounter_id, user_id, medications)

@spec order(String.t(), String.t() | nil, String.t(), [medication_order_input()]) ::
  {:ok, Owat.Model.MedicationRequest.t()}
  | {:error, :not_found | Owat.CrudUtils.validation_error()}

Directly order a medication, skipping the proposal part.

reject_request(id, encounter_id, user_id, reason)

@spec reject_request(String.t(), String.t() | nil, String.t(), String.t()) ::
  {:ok, Owat.Model.MedicationRequest.t()}
  | {:error, :not_found | :invalid_status | Owat.CrudUtils.validation_error()}

Reject a request proposal with a given reason.

request(person_id, encounter_id, user_id, medications)

@spec request(String.t(), String.t(), String.t(), [medication_request_input()]) ::
  {:ok, Owat.Model.MedicationRequest.t()}
  | {:error, :not_found | Owat.CrudUtils.validation_error()}

Request a new medication.

This does not issue a prescription, but merely signals a users proposal to prescribe a mediation to a person.

rescind_request(id, person_id, user_id)

transact(changesets)