Owat.Meilisearch (PVE v0.0.1)

Provides functions to interact with the Meilisearch API.

Functions

create_address_index()

init()

@spec init() :: :ok | {:error, Meilisearch.Error.t(), String.t()}

Initializes the Meilisearch client and creates the patient index if it does not exist.

Returns an error if an error other than the missing index occurs.

insert_addresses(addresses)

search_address(query, filter)

@spec search_address(String.t() | nil,
  country: String.t(),
  postal_code: String.t(),
  city: String.t(),
  delivery_locality: String.t(),
  municipality: String.t(),
  street_name: String.t(),
  number: String.t()
) :: {:ok, [%{required(String.t()) => String.t()}]} | {:error, term()}

search_address(query, attributes, filter)

@spec search_address(String.t() | nil, [String.t()],
  country: String.t(),
  postal_code: String.t(),
  city: String.t(),
  delivery_locality: String.t(),
  municipality: String.t(),
  street_name: String.t(),
  number: String.t()
) :: {:ok, [%{required(String.t()) => String.t()}]} | {:error, term()}

search_medication(query)

search_patients(query)

@spec search_patients(String.t()) ::
  {:ok, [Owat.Model.Person.t()]}
  | {:error, :invalid_query | :unavailable | :unknown}

Searches the patient index for the given query.

Returns a list of up to 20 patients or an empty list if no patients were found or the query was invalid.

sync_medications(progress_reporter)

sync_patients(progress_reporter)

@spec sync_patients((integer() -> any())) :: :ok

Deletes all documents from the patient index and inserts all patients from the database.

update_patient(person)

@spec update_patient(Owat.Model.Person.t()) :: any()

Updates the patient index with the given people.

Existing entries will be replaced, if a value in the new entry is nil or missing, the previous value will be lost.

wait_for_tasks(client, options \\ [])