Owat.Model (PVE v0.0.1)

Base functionality for models that get sent over the websocket connection.

When used this module will import the datatype which will do the following when invoked:

Generate a datamap function that consumes the struct of the caller and produces a stripped map according to the given rules in the do block

Generate a zod schema file for each datatype that is used in the module

Usage

Generally a list of :-> expressions is given that specify which keys the resulting map shall have.

datatype :fn_name do
  :resulting_field_name ->
    :struct_field
end

This will produce the function fn_name/1 like this:

def fn_name(internal_name) do
  %{
    resulting_field_name: internal_name.struct_field
  }
end

Complex Sub-Fields

When the model has complex sub-fields, a parenthesized subexpression can be given. This subexpression gives the struct field name and the corresponding type module. These fields are interpreted as optional if not specified otherwise.

datatype :fn_name do
  :resulting_field_name ->
    :struct_field :: Module.datatype_fn()
end

# ===>

def fn_name(internal_name) do
  %{
    resulting_field_name: Module.datatype_fn(internal_name.struct_field)
  }
end

When additional values need to be passed the following method is used to pass them:

datatype :fn_name do
  :resulting_field_name ->
    :struct_field :: Module.datatype_fn("supplemental", "arguments")
end

# ===>

def fn_name(internal_name) do
  %{
    resulting_field_name: Module.datatype_fn(struct.struct_field, ["supplemental", "arguments"])
  }
end

This method of producing a value for a subfield by the subfields module can also be used when in need to produce lists. In this case the module atom or module invocation has to be placed in list brackets. This is also required in order to ensure that the zod type validation adapts to an array type.

datatype :fn_name do
  :without_arguments -> :field_a :: [Module.datatype_fn()]
  :with_arguments -> :field_b :: [Module.datatype_fn("supplemental", "arguments")]
end

# ===>

def fn_name(internal_name) do
  %{
    without_arguments: Enum.map(struct.field_a, &Module.datatype_fn(&1))
    with_arguments: Enum.map(struct.field_b, &Module.get_list(&1, ["supplemental", "arguments"]))
  }
end

Type overrides and optionals

When a field's type cannot be automatically inferred or the type does not support zod typing, a type override can be specified.

datatype :fn_name do
  :resulting_field_name, TypeOverride ->
    :struct_field
end

When a field is optional the left side of the arrow shall be wrapped in a call to optional:

datatype :fn_name do
  optional(:simple) ->
    :struct_field_a

  optional(:with_type_override, TypeOverride) ->
    :struct_field_b
end

Precomputed values

Values of the resulting data map can also be arbitrary code. This is done by wrapping the expression with the pin operator.

datatype :fn_name do
  :result -> ^Module.call()
end

The struct is generally invisible. When the value depends on the struct it has to be made visible. This is done with the :binds option that also allows precomputing values to be used in multiple fields.

datatype :fn_name, binds: struct do
  :result -> ^Module.call(struct)
end

datatype :fn_name, binds:
        (struct ->
          variable = Module.call(struct)
        ) do
  :a -> ^variable.a
  :b -> ^variable.b
end

Functions

datatype(name, opts \\ [], list)
(macro)

newest(list)

newest(list, key)