Owat.Model.LifeInfo (PVE v0.0.1)

Record of information related to a person's life.

The following records types exist:

:birth

:death

:citizen

Birth Record

A birth record is of kind = :birth and has a date. A birth record without date does not provide any information, since it is assumed, that any patient's life has begun. A birth record neither has the confirmed flag nor string info.

Death Record

A death record is of kind = :death and may have a date.

The existence of a record without a date indicates that a person has died without the date of death being known.

The death record may be confirmed by setting the flag true or explicitly unconfirmed by setting the flag false. It may also be of unknown confirmation state when the flag is nil.

A death record does not have string info.

Citizenship Record

A citizenship record is of kind = :citizen.

A citizenship record has neither a date nor a confirmed flag.

The citizenship is stored as text info containing the ISO 2-letter country code.

Types

birth_record()

@type birth_record() :: %Owat.Model.LifeInfo{
  __meta__: term(),
  confirmed: nil,
  id: term(),
  info: nil,
  inserted_at: DateTime.t(),
  kind: :birth,
  lower_bound: Date.t(),
  person: Owat.Repo.relation(Owat.Model.Person.t()),
  person_id: String.t(),
  source: Owat.Terminology.Code.t(),
  superseded_at: DateTime.t() | nil,
  updated_at: DateTime.t(),
  upper_bound: Date.t()
}

citizenship_record()

@type citizenship_record() :: %Owat.Model.LifeInfo{
  __meta__: term(),
  confirmed: nil,
  id: term(),
  info: String.t(),
  inserted_at: DateTime.t(),
  kind: :citizen,
  lower_bound: nil,
  person: Owat.Repo.relation(Owat.Model.Person.t()),
  person_id: String.t(),
  source: Owat.Terminology.Code.t(),
  superseded_at: DateTime.t() | nil,
  updated_at: DateTime.t(),
  upper_bound: nil
}

death_record()

@type death_record() :: %Owat.Model.LifeInfo{
  __meta__: term(),
  confirmed: boolean() | nil,
  id: term(),
  info: nil,
  inserted_at: DateTime.t(),
  kind: :death,
  lower_bound: Date.t() | nil,
  person: Owat.Repo.relation(Owat.Model.Person.t()),
  person_id: String.t(),
  source: Owat.Terminology.Code.t(),
  superseded_at: DateTime.t() | nil,
  updated_at: DateTime.t(),
  upper_bound: Date.t() | nil
}

t()

@type t() :: birth_record() | death_record() | citizenship_record()

Functions

get_value(list, kind)

select(list, kind)