Owat.Model.Task (PVE v0.0.1)

Task Model

Types

t()

@type t() :: %Owat.Model.Task{
  __meta__: term(),
  assignee: term(),
  assignee_id: term(),
  code: String.t(),
  completer: Owat.Model.User.t() | nil,
  completer_id: term(),
  description: String.t() | nil,
  done_at: DateTime.t() | nil,
  encounter: Owat.Model.Encounter.t() | nil,
  encounter_id: term(),
  id: term(),
  inserted_at: term(),
  intent: Owat.Terminology.Code.t(),
  medication_request_reception: term(),
  medication_request_reception_id: term(),
  parent: t() | nil,
  parent_id: term(),
  patient_instruction: String.t() | nil,
  patient_visible: boolean(),
  priority: Owat.Terminology.Code.t(),
  requester: Owat.Model.User.t() | nil,
  requester_id: term(),
  sort_order: float(),
  started_at: DateTime.t() | nil,
  status: Owat.Terminology.Code.t(),
  subject: Owat.Model.Person.t() | nil,
  subject_id: term(),
  subtasks: [t()],
  system_task: boolean(),
  updated_at: term(),
  working_set: map() | nil
}

Functions

completed_status?(code)

Determines if a given terminology code represents either a done or an error task status.

completed_status_codes()

datamap(data)

done_status?(code)

@spec done_status?(Owat.Terminology.Code.t()) :: boolean()

Determines if a given terminology code represents a done task status.

encounter(data)

error_status?(code)

@spec error_status?(Owat.Terminology.Code.t()) :: boolean()

Determines if a given terminology code represents an error task status.

make_payload(arg1)

ongoing_status?(code)

@spec ongoing_status?(Owat.Terminology.Code.t()) :: boolean()

Determines if a given terminology code represents an ongoing task status.

pending_status?(code)

@spec pending_status?(Owat.Terminology.Code.t()) :: boolean()

Determines if a given terminology code represents a pending task status.

sorted(tasks)