Owat.Permissions (PVE v0.0.1)

This module exposes functions for capabilities-based permissions and also acts as plug for enforcing these.

Available Permissions

These are the currently available capabilities:

Technical Settings (:technical)

Integrations (Planery, Windoc, Devices, ...)

Call Display

Organizational Settings (:organizational)

Permissions

Organization OID

E-Card / GINS Settings

Board Columns

Administrative Settings (:administrative)

Healthcare Services

Scheduling

Opening Hours

Task Templates

General Access (:general_access)

Enables functions on patients that are not the users own file.

Enables booking more than 1 month ahead

Disables checking for minimum booking distance

Prescribing (:prescribing)

Direct ordering of medications

Accepting medication requests

Using this module as plug

This module can be used in two ways as authentication plug.

As simple permission gate:

plug Owat.Permissions, permission: :technical

This simply checks if the logged-in user has the specified permission.

As person authorization:

plug Owat.Permissions, person_id_key: "person_id"

This will take the "person_id" key from the Conn's params and check if the logged-in user has this ID assigned or if the user has the :general_access permission.

Types

permission()

@type permission() ::
  :technical
  | :organizational
  | :administrative
  | :general_access
  | :prescribing

Functions

all()

@spec all() :: [permission()]

Returns a list of all available permission atoms.

authorize(socket_or_conn, permissions_atom)

@spec authorize(Plug.Conn.t() | Phoenix.Socket.t(), permission()) ::
  :ok | {:error, :permission_insufficient}

@spec authorize(Plug.Conn.t() | Phoenix.Socket.t(), String.t()) ::
  :ok | {:error, :permission_insufficient | :not_found}

Checks if a given user has permission.

The user can be given by an authenticated Plug.Conn or Phoenix.Socket struct. The permission is either given as permission atom or the ID of a person record.

This function returns :ok when the user has permission, or {:error, reason} otherwise. reason is :permission_insufficient when the user does not have permission. When the user has :general_access permission and a non-existent person ID is given :not_found will be returned as reason.

Example Usage (case-style)

def route(conn, _params) do
  case Owat.Permissions.authorize(conn, :technical) do
    :ok ->
      perform_technical_action(conn)

    {:error, :permission_insufficient} ->
      conn
      |> put_status(403)
      |> json(%{message: "forbidden"})
  end
end

Example Usage (with-style)

def route(conn, _params) do
  with :ok <- Owat.Permissions.authorize(conn, :technical),
       :ok <- other_preconditions(conn) do
      perform_technical_action(conn)
  else
    {:error, :permission_insufficient} ->
      conn
      |> put_status(403)
      |> json(%{message: "forbidden"})
  end
end

employees()