Owat.Roster.Provider behaviour (PVE v0.0.1)

A behaviour for Roster Providers.

The Roster Provider is a connector-type module that handles communications with the upstream roster source. The roster provider module is thought to not have to manage state, but it is expected that upstream communications is stateful. For this reason every function takes as last parameter the current state and returns a tuple where its last element is the new state.

The init/0 callback is used to obtain the initial state.

Types

state()

@type state() :: term()

The state of the roster provider.

Can be any term. Will be nil initially.

Callbacks

init()

@callback init() :: state()

Returns the initial state.

shifts!(t, state)

@callback shifts!(Date.t(), state()) :: {[Owat.Roster.ShiftInfo.t()], state()}

Returns the shift information for the given date.