Owat.Support (PVE v0.0.1)

Various utility functions

Functions

boolean_env(key, default \\ false)

@spec boolean_env(String.t(), boolean()) :: boolean()

Obtains a boolean environment variable.

The following values are considered false: "", "f", "false", "n", "no", "0" Any other value is considered true.

If the key is not set, the default argument is returned.

e164(number)

@spec e164(String.t()) :: {:ok, String.t()} | {:error, String.t()}

Parse a phone number and format it in E.164 format.

Returns either an ok-tuple with a valid E.164 formatted phone number or an error-tuple with a human-readable error message string.

exclude_keep(changes, nonce \\ :keep)

found(enumerable, fun)

@spec found(Enum.t(), (Enum.element() -> any())) ::
  {:ok, Enum.element()} | {:error, :not_found}

The same as Enum.find/2 but returns a tuple of {:ok, element} when found and {:error, :not_found} otherwise.

get_or_keep_fields(data, fields, nonce \\ :keep)

@spec get_or_keep_fields(%{required(String.t()) => term()}, [atom()], atom()) :: %{
  required(atom()) => term()
}

Extracts given fields from a map.

Returns a map with the specified atoms as keys. A tuple of atoms {old_key, new_key} can be used to rename keys. Any key atom that has a corresponding string key in the original map will result in the original value. All keys of the original map that are either not mentioned or not strings will be discarded. Keys not present in the original map will get the nonce value assigned in the result.

Examples

iex> Support.get_or_keep_fields(%{"take" => 1, "also" => nil, "drop" => "me"}, [:take, :also, :some])
%{take: 1, also: nil, some: :keep}

iex> Support.get_or_keep_fields(%{"take" => 1, "also" => nil, "drop" => "me"}, [:take, {:also, :renamed}, :some])
%{take: 1, renamed: nil, some: :keep}

iex> Support.get_or_keep_fields(%{:key => "foo"}, [:key])
%{key: :keep}

iex> Support.get_or_keep_fields(%{}, [:key], :foo)
%{key: :foo}

html_escape(other)

@spec html_escape(iodata()) :: iodata()

Escapes HTML entities in the given iodata.

integer_env(key, default \\ nil)

@spec integer_env(String.t(), integer() | nil) :: integer()

Obtains an integer environment variable.

If the key is not set or cannot be parsed, the default argument is returned.

merge_maps(map \\ %{}, list_of_maps)

@spec merge_maps(map(), [map()]) :: map()

Merges many maps.

Maps are merged head-first. With keys in multiple maps the last one prevails. This function performs shallow merges.

Examples

iex> Support.merge_maps([%{a: "a"}, %{b: "b"}, %{c: "c"}])
%{a: "a", b: "b", c: "c"}

iex> Support.merge_maps([%{a: "a", b: "b"}, %{b: "c", c: "c"}])
%{a: "a", b: "c", c: "c"}

iex> Support.merge_maps([%{a: %{x: "x"}}, %{a: %{y: "y"}}])
%{a: %{y: "y"}}

random_ref()

@spec random_ref() :: String.t()

Generates a random reference string.

strip(map, strip_all? \\ false)

@spec strip(%{optional(term()) => term()}, boolean()) ::
  %{optional(term()) => term()} | nil

Recursively strips all nil values from a bare map.

Only bare maps are stripped. Empty maps will be removed except the top-level map. The top-level map can also be removed when settings strip_all? to true. Structs are never touched.

Examples

iex> Support.strip(%{keep: 1, drop: nil})
%{keep: 1}

iex> Support.strip(%{:only => nil, "keys" => nil})
%{}

iex> Support.strip(%{:only => nil, "keys" => nil}, true)
nil

iex> Support.strip(%{recurse: %{keep: 1, drop: nil}, empty: %{drop: nil}})
%{recurse: %{keep: 1}}

iex> Support.strip(%{structs: %Foo{keeps: "all", fields: nil}})
%{structs: %Foo{keeps: "all", fields: nil}}