Owat.Support.DateTime (PVE v0.0.1)

Various utility functions for DateTime

Types

partial_date()

@type partial_date() ::
  {:calendar.year()} | {:calendar.year(), :calendar.month()} | :calendar.date()

A partial date

Functions

at(date, time)

@spec at(Date.t(), Time.t()) :: DateTime.t()

end_of_day(date)

@spec end_of_day(Date.t()) :: DateTime.t()

Returns the end of the day of the given date as datetime.

holiday?(date)

@spec holiday?(Date.t() | String.t()) :: boolean()

Checks if the given date is a holiday.

local_date(datetime)

@spec local_date(DateTime.t()) :: Date.t()

Shifts a DateTime to the local timezone and extracts its date information.

now()

@spec now() :: DateTime.t()

Returns the current time truncated to the given precision.

now(precision)

@spec now(precision :: :microsecond | :millisecond | :second) :: DateTime.t()

partial_date_to_range(erl)

@spec partial_date_to_range(partial_date()) :: {Date.t(), Date.t()}

Converts a partial_date/0 to a range of dates.

start_of_day(date)

@spec start_of_day(Date.t()) :: DateTime.t()

Returns the start of the day of the given date as datetime.

today()

@spec today() :: Date.t()

Returns today's date in the configured timezone.

xml_partial_date(arg1)

@spec xml_partial_date(String.t()) :: partial_date() | :error

Parses a possibly-partial date string.

Supported formats are:

YYYYMMDD

YYYYMM

YYYY

The returned Date Struct has