Owat.Support.Presence (PVE v0.0.1)

A module to manage presence state.

This module contains the actionable portion of presence functionality. It must be used with a state-carrying facility, such as a GenServer.

Method

This module has functions to track resources in a presence channel. It only tracks the IDs. When an ID is added, removed or changed it will invoke a fetcher function to obtain the payload to be sent to the client.

Usage

To use the functions in this module a state value has to be generated. To generate this value you have to supply the channel name and fetcher function. Optionally initial data may be passed to init.

Owat.Support.Presence.init("channel_name", &fetch/1)
# or
Owat.Support.Presence.init("channel_name", &fetch/1, [])

To update the state use put/2 and delete/2.

Local mode

This module broadcasts the diff-updates on the given channel or only on a specific socket, previously called local-mode. To enable local mode, pass the socket instead of a channel name to the init/3 function.

Owat.Support.Presence.init(socket, &fetch/1, [])

The init function itself does not send data to the client.

Fetcher Function

The fetcher function must produce the payload objects to send to the client. It is given a list of IDs and returns a map of payload objects that are keyed with their respective IDs. Returning an enumerable that can be directly converted to this map is also acceptable.

Types

data()

@type data() :: %{id: String.t()} | String.t()

Data to be tracked in the presence state.

Any object tracked in the presence state must have a unique ID and this ID must be in the :id field of the map. The presence state only tracks the contained IDs. An ID is transformed be the fetcher function to the payload to be transmitted. Any function receiving data information may also only receive the ID of the object to be tracked.

fetcher()

@type fetcher() :: ([String.t()] -> %{required(String.t()) => map()})

A function that takes a list of IDs and produces a map of %{id => data} where data is the payload to be sent to the clients.

state()

@opaque state()

The current presence state.

An instance of this type contains all necessary information to perform actions with this module. The state data is meant to be kept in a state-carrying facility such as a GenServer.

All functions (except init/3 of course) take the state parameter as first argument. Modifying functions return the new state. The state shall not be modified by other means than this module.

Functions

delete(arg, obj)

@spec delete(state(), data() | [data()]) :: state()

Deletes an object or list of objects from the state if present.

init(channel_name_or_socket, fetcher, initial_data \\ [])

@spec init(String.t() | Phoenix.Socket.t(), fetcher(), [data()]) :: state()

Produces an empty presence state.

The first parameter may either be a channel name, where all updates are broadcasted to or a socket when the presence state is local to a single socket (local mode).

The second parameter is the fetcher function. Since the state only tracks the IDs of the objects, the fetcher function is necessary to fetch the payload data to be sent to the clients on change.

Optionally the initial data may be supplied as third argument.

This function does not send data to clients.

keys(arg)

@spec keys(state()) :: [String.t()]

Obtains the IDs of the resources currently present.

push(arg, socket)

@spec push(state(), Phoenix.Socket.t()) :: :ok

Pushes the presence state to a socket.

put(arg, obj)

@spec put(state(), data() | [data()]) :: state()

Adds an object or list of objects to the state, replacing any previous entry with the same id.

put_direct(arg, id, obj)

@spec put_direct(state(), String.t(), map()) :: state()

Adds an object or list of objects to the state, replacing any previous entry with the same id.

This function, in contrast to put/2 does not invoke the fetcher and does therefore require to have already the prepared datamap passed.

reset(arg, list)

@spec reset(state(), [data()]) :: state()

Resets the state to contain only the given elements.

Beware

This action will not send updates for any elements that were previously in the state. It will only join elements not present in the current state and leave elements not present in the argument.

reset_direct(arg, list)

@spec reset_direct(state(), map()) :: state()

Resets the state to contain only the given elements.

In contrast to reset/2, this function does not invoke the fetcher and rejoins all elements regardless of the previous state. This function expects already prepared datamaps to be passed.