Owat.Support.Sorting (PVE v0.0.1)

This module provides helper functions for sorting models.

Requirements

Every model must be equipped with a :sort_order field.

After any operation that updates the database state, the sanitizing functions must be called.

Functions

before(queryable, id)

@spec before(Ecto.Queryable.t(), String.t() | nil) :: integer()

Calculates the sort_order before a given entry by id.

If an entry is given, the returned sort_order value is directly preceding the given entry. If nil is given, the returned sort_order value is last.

put_sort_order(data, queryable)

sanitize!(model, where \\ [])

@spec sanitize!(module(), Keyword.t()) :: :ok

Runs the sanitation routine to respace all sort_order fields.

Optionally, a binding can be given to filter the sanitation to a specific subset of the model.