Owat.WorkIncapacity (PVE v0.0.1)

Business logic to work with work-incapacity conditions.

Functions

delete(cid)

@spec delete(String.t()) ::
  {:ok, :ok}
  | {:error, :not_found | :delete_not_allowed | Owat.Gin.SS12.fault_info()}

Deletes a work-incapacity condition given its ID.

This will also cancel it on the e-Card System.

fetch(encounter)

@spec fetch(Owat.Model.Encounter.t()) :: [Owat.Model.Condition.t() | term()]

Tries to update all conditions from the SS12.

Since conditions are linked to an encounter this function must be called with an encounter to which new conditions are linked to.

for(pid)

@spec for(String.t()) :: [Owat.Model.Condition.t()]

Obtains all work-incapacity conditions for a given person's ID.

get(id)

@spec get(String.t() | [String.t()]) ::
  Owat.Model.Condition.t() | [Owat.Model.Condition.t()] | nil

Obtains one or many work-incapacity conditions by its IDs.

Returns nil if the condition was not found or if the condition with this ID is not a work-incapacity condition.

When the argument is a list, the return type is also a list.

get_with(id)

@spec get_with(String.t() | [String.t()]) ::
  {:ok, Owat.Model.Condition.t() | [Owat.Model.Condition.t()]}
  | {:error, :not_found}

The same as get/1 but returns an ok-tuple when found or an error-tuple when not found.

new_condition(encounter, source, author_id, clinical_status, verification_status, attrs \\ %{})

@spec new_condition(
  Owat.Model.Encounter.t(),
  String.t(),
  String.t(),
  String.t(),
  String.t(),
  map()
) ::
  {:ok, Owat.Model.Condition.t()} | {:error, Ecto.Changeset.t()}

Create a new work-incapacity condition.

submit(cid)

@spec submit(String.t()) ::
  {:ok, Owat.Model.Condition.t()}
  | {:error,
     :no_ssn
     | :no_diagnosis
     | :no_foreign_id
     | :no_active_consult
     | :consult_has_no_voucher
     | :person_must_have_arrived
     | Owat.Gin.SS12.fault_info()}

Submits a condition to the SS12.

update(cid, data, author_id)

@spec update(String.t(), map(), String.t() | nil) ::
  {:ok, Owat.Model.Condition.t()}
  | {:error, :not_found | :edit_not_allowed | Ecto.Changeset.t()}

From the condition with the given ID update the given data.