Owat.Workflow.OngoingEncounters (PVE v0.0.1)

GenServer to handle ongoing encounter presence.

This GenServer keeps the list of currently ongoing encounters in memory. Changes made directly in the DB are scanned with a message timeout every 60s of inactivity. A scan can also be triggered manually by using rescan/0.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

list()

@spec list() :: [String.t()]

Get the list of ongoing encounter IDs.

push(socket)

@spec push(Phoenix.Socket.t()) :: {:noreply, Phoenix.Socket.t()}

Pushes the current ongoing encounter state to the given socket.

rescan()

@spec rescan() :: :ok

Rescan the currently ongoing encounters from the DB.

start_link(opts)

sync(encounter)

@spec sync(Owat.Model.Encounter.t()) :: :ok

Updates the internal state to reflect the status of the given encounter.

Will add the encounter ID to the state if the status is considered ongoing and not yet present in the state. Will delete the encounter ID if the status is not considered ongoing and the ID was previously present in the state.