Owat.Workflow.Tasks (PVE v0.0.1)

Functions for handling tasks

Types

task_option()

@type task_option() ::
  {:code, String.t() | nil}
  | {:parent_id, String.t() | nil}
  | {:subject_id, String.t() | nil}
  | {:encounter_id, String.t() | nil}
  | {:assignee_id, String.t() | nil}
  | {:medication_request_reception_id, String.t() | nil}
  | {:description, String.t() | nil}
  | {:intent, integer() | nil}
  | {:priority, String.t() | nil}
  | {:patient_visible, boolean() | nil}
  | {:patient_instruction, String.t() | nil}
  | {:before, String.t() | nil}

Functions

by_ids(id_list)

create_task(requester_id, opts)

@spec create_task(String.t(), [task_option()]) ::
  {:ok, Owat.Model.Task.t()}
  | {:error,
     :not_found
     | :no_subject_encounter_match
     | :no_description
     | Owat.CrudUtils.validation_error()}

Create a task on an encounter.

The status will be set to requested and the requester will be set to the user making the request. The sort_order will be set according to the before value if it matches another task_id on the encounter, or to the last position in the encounter. If code is provided, it has to match an existing task template. patient_instruction cannot be provided without setting patient_visible to true.

delete_task(id)

@spec delete_task(String.t()) :: {:ok, Owat.Model.Task.t()} | {:error, term()}

Delete a task.

maybe_broadcast(arg)

maybe_delete(struct_or_changeset)

maybe_insert(changeset, preload_opts \\ [])

maybe_transact(transaction)

maybe_transact(transaction, change_to_broadcast)

maybe_update(changeset)

maybe_update(changeset, preload_opts)

maybe_update(state, changes, preload_opts)

transact(changesets)

update_task(id, user_id, data)

@spec update_task(String.t(), String.t(), %{
  optional(:status) => String.t() | nil,
  optional(:before) => integer() | nil
}) :: {:ok, Owat.Model.Task.t()} | {:error, term()}

Update a task's status and sort_order.

If the status is set to a completed state, the completer_id will be set to the user_id of the user making the request.