Owat.ZodTransformer (PVE v0.0.1)

Transforms an OWAT AST into a Zod module.

Types

array()

@type array() :: {:array, base_type() | schema()}

An array in zod

assign()

@type assign() :: {:assign, String.t(), base_type() | schema()}

A typescript assign statement

(Useful for referencing a schema in a different module)

ast()

@type ast() :: [import() | enum() | object() | infer() | interface()]

A Zod AST

base_type()

@type base_type() ::
  :string
  | :number
  | :integer
  | :boolean
  | :date
  | :datetime
  | :record
  | :any
  | :void
  | nil
  | array()
  | literal()

A string in zod

enum()

@type enum() :: {:enum, String.t(), [String.t()]}

An enum consisting of the name of the enum and a list of the enum's values.

enum_field()

@type enum_field() :: {:enum, [String.t()]}

An enum field of an object in zod

extended()

@type extended() ::
  {:extended, String.t(), schema(), [field() | object()], [modifier()]}

An extended object in zod

field()

@type field() :: {String.t(), base_type() | enum_field() | schema(), [modifier()]}

An object field in zod

import()

@type import() ::
  {:import, String.t(),
   [
     {String.t()}
     | {String.t(), String.t()}
     | {:type, String.t()}
     | {:type, String.t(), String.t()}
   ]}

An import statement containing the name of the module to import and the names of the exports or types to import.

represents a type import. {String.t, String.t} represents an import renamed from the original name to a new name. {:type, String.t, String.t} represents a type import renamed from the original name to a new name.

infer()

@type infer() :: {:infer, String.t(), schema()}

An inferred type in zod

interface()

@type interface() ::
  {:interface, String.t(), [interface_field()], String.t()}
  | {:interface, String.t(), [interface_field()]}

A typescript interface defintion

The last argument in the 4-tuple is a referene to the parent interface

interface_field()

@type interface_field() ::
  {String.t(), String.t() | :void | schema() | array() | [interface_field()]}

An interface field

literal()

@type literal() :: {:literal, String.t()}

A literal in zod

modifier()

@type modifier() :: :optional | :nullable | :strict | :partial

A modifier in zod

object()

@type object() :: {:object, String.t(), [field() | object()], [modifier()]}

An object in zod

schema()

@type schema() :: {:schema, String.t()}

A reference to a previously defined or imported schema

union()

@type union() :: {:union, String.t(), [base_type() | schema()]}

A union in zod

Functions

base_name(module)

@spec base_name(module()) :: String.t()

Produce the base name of a module.

This is the last part of the module atom.

Examples

iex> Zod.base_name(Foo.Bar.Baz)
"Baz"

dcfirst(arg)

embedded_name(module)

mapping_name(module, datatype_fn)

schema_name(name)

@spec schema_name(module() | String.t()) :: String.t()

Produce the schema name of a module.

This is the base name, lowercase first character, with "Schema" as suffix.

Examples

iex> Zod.schema_name(Foo.Bar.Baz)
"bazSchema"

transform(list)

@spec transform(import() | base_type() | enum() | object() | infer() | interface()) ::
  iodata()

Takes an AST node and returns a string containing the Zod schema definition as typescript code.

transform_ast(ast)

@spec transform_ast(ast()) :: String.t()

Takes an AST and returns a string containing the Zod schema definition as typescript code.

type_name(module, datatype_fn)

ucfirst(arg)

write_module(name, ast)

@spec write_module(String.t(), ast()) :: String.t()

Transforms an ast to zod schema definitions and writes them to a file using the given name.