OwatWeb.AuthController (PVE v0.0.1)

This controller handles all authentication aspects.

The authentication is generally divided into two parts:

OIDC general authentication

Token claims

For all restricted services a OIDC general authentication must be obtained. This is done by following the OIDC auth workflow for any available provider. Currently there are the following providers:

azure - Azure

eid - ID Austria

The OIDC general authentication will issue a session cookie which has all authentication information attached. This cookie has to be transmitted on every call to restricted services.

OIDC General Authentication Flow

When the user agent navigates to /auth/:provider/redirect it will be issued a 302 redirect response with all the necessary information. The redirect will directly lead to the authentication page of the selected provider.

After successful authentication the user agent is redirected by the IdP to the callback url. Then the session is authenticated and can be used to access restricted services.

Token claims

To access websocket connections a specific token claim has to be issued. This is to mitigate CSWSH attacks.

A token claim for websocket connections can be obtained at /api/self/websocket-token. The obtained token must then be passed as GET parameter in the websocket URL.

Functions

callback(conn, params)

csrf_token(conn, params)

default_login(conn, params)

logout(conn, params)

request(conn, params)

websocket_token(conn, params)

Issues a websocket token claim to an authenticated user.