OwatWeb.GlobalChannel (PVE v0.0.1)

This module handles the global topic.

Any socket is authorized for this topic.

The following events exist:

message

Unmatched events

Any unmatched event gets an unicast reply with status error informing the sender of the non-existence of this event type.

The response always contains the type: "unmatched_event" and the information which event was received.

%{
  type: "unmatched_event",
  event: event
}

Bidirectional Events

These events can be sent to the server and can be received unsolicited from the server.

message Event

Payload	SmallMessage

Permissions	Anyone

The message event will be broadcasted as is to all subscribers.

Upstream Events

The following events will never be sent by the server.

search Event

This event is for searching in the cores search index.

Payload	SearchQuery

Permissions	Anyone

The sender will get a reply with the search results.

OK-Reply

The reply with status ok contains the search results.

{
  "estimatedTotalHits": 0,
  "hits": [ PersonSummary ],
  "limit": 0,
  "offset": 0
}

Data types

PersonSummary

{
  "address": "string",
  "emails": [ "string:email" ],
  "family": "string",
  "given": "string",
  "id": 0,
  "phones": [ "string:e164" ],
  "ssn": "string"
}

SmallMessage

All fields are required.

{
  "message": "string",
}

Functions

child_spec(init_arg)

handle_in(event, params, socket)

Handles events sent to this channel.

join(binary, msg, socket)

Joins a given socket into the global channel.

start_link(triplet)