OwatWeb.HtmlSafeSerializer (PVE v0.0.1)

HTML-safe Websocket Serializer for Phoenix

Usage

This is a wrapping serializer and cannot be used on its own. A concrete implementation with an existing serializer must be created first.

defmodule MySerializer do
  use Elixir.OwatWeb.HtmlSafeSerializer, from: ExistingSerializer
end

How it works

This module wraps the given Phoenix.Socket.Serializer and preprocesses messages when encoding data. When a message is received it is sanitized and then passed to the used base-serializer.

Sanitation occurs based on the following rules:

Tuples of {:safe, data} result directly in data without any processing.

Atoms stay untouched.

Binaries are passed through Owat.Support.html_escape/1.

Lists are sanitized element-wise.

Structs get converted to maps, then sanitized as maps and then reformed to structs.

Maps are sanitized by key and value using io-unsafe sanitation.

Any other type including numerics and tuples cannot be sanitized. This further restricts iodata sanitation, since iodata containing numeric codepoints cannot be sanitized.

io-unsafe sanitation works the same as normal sanitation except with those two key differences:

Numerics stay untouched.

Tuples are io-unsafe sanitized element-wise.

Functions

sanitize(message)